==> 07_references.Rmd <==
`r if (knitr::is_html_output()) '# References {-}'`


```{r prepare-data-package, echo = FALSE}
zip_file <- "data/data.zip"
files <- c("ebd_woothr_june_bcr27.txt", 
           "ebd_june_bcr27_sampling.txt",
           "ebd_woothr_june_bcr27_zf.csv",
           "gis-data.gpkg",
           "mcd12q1_classes.csv",
           "modis_pland_location-year.csv",
           "modis_pland_prediction-surface.csv",
           "prediction-surface.tif")
unlink(zip_file)
zip(zip_file, file.path("data", files))
```
